+++ src/main/preload.ts
@@ -95,7 +95,25 @@ contextBridge.exposeInMainWorld('electronAPI', {
     return () => {
       ipcRenderer.removeListener(channel, listener);
     };
-  }
+  },
+
+  // Navigation
+  updateNavigationState: (canGoBack: boolean, canGoForward: boolean) => 
+    ipcRenderer.send('navigation:updateState', canGoBack, canGoForward),
+  
+  onNavigateBack: (callback: () => void) => {
+    ipcRenderer.on('navigation:back', callback);
+    return () => {
+      ipcRenderer.removeListener('navigation:back', callback);
+    };
+  },
+
+  onNavigateForward: (callback: () => void) => {
+    ipcRenderer.on('navigation:forward', callback);
+    return () => {
+      ipcRenderer.removeListener('navigation:forward', callback);
+    };
+  },
 });
 
 declare global {
@@ -142,6 +160,9 @@ declare global {
       onTimeTrackingChanged: (callback: () => void) => () => void;
       onSetTheme: (callback: (theme: string) => void) => () => void;
       onUpdateStatus: (callback: (payload: UpdateStatusPayload) => void) => () => void;
+      updateNavigationState: (canGoBack: boolean, canGoForward: boolean) => void;
+      onNavigateBack: (callback: () => void) => () => void;
+      onNavigateForward: (callback: () => void) => () => void;
     };
   }
 }